adding script to embed svg to markdown
this script is taken from https://stackoverflow.com/questions/44139441/inlining-svg-image-in-a-markdown-cell and I personally use it to embed svg into Joplin which is my preferred notes application on all my operating systems. I have not modified the scrip other than inserting a few print arguments in it a rearranging the code a bit.